+++ src/main/java/com/luxes/dev/expensetracker/model/Task.java
@@ -6,12 +6,13 @@ import java.time.LocalDate;
 
 
 public record Task(
-        long id,
+        int id,
         @NotBlank(message = "Description should not be blank.")
         String description,
         @NotBlank(message = "Category should not be blank.")
         String category,
         LocalDate dateCreation,
         double cost
-) {}
+) {
+}
